// frontend/src/app/core/services/api-client/model/itemRutinaReq.ts
/**
 * MyFitnessApp
 * A REST api for exercise and workout tracking
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SerieReq } from './serieReq';


export interface ItemRutinaReq { 
    ejercicioId: number;
    descansoEnSeg: number;
    nota: string;
    series: Array<SerieReq>;
}