// frontend/src/app/core/services/api-client/model/serieReq.ts
/**
 * MyFitnessApp
 * A REST api for exercise and workout tracking
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SerieReq { 
    reps?: number;
    pesoEnKg?: number;
    tiempoEnSeg?: number;
    distancia?: number;
}